Completed giphyAPI.py and giphyCLI.py files / classes. Added bash script to allow me to run my app with the command gif. Made some minor fixes here and there.
